Process multiple runs
Process reads the same config as duetbench, thus separate the config
parser and use in both duetbench and process renaissance.

Process renaissance script relies on the way duetbench stores results
i.e. single directory with multiple properly named files, this should be
extensible for other types of benchmarks as well since the naming
scheme will be kept even for multiple files specified in config and the
processing script has access to this config.