Initial Asynchronous-Duet workflow (#1)
Add support for Renaissance suite.

YAML config deprecates all the run scripts, therefore single Dockerfile
suffices for Renaissance benchmark. With that get rid of old alpine
version and all the scripts as well as original shell duetbench script.

Add option to run duet multiple times with different ordering of A and B.

Process reads the same config as duetbench, thus separate the config
parser and use in both duetbench and process renaissance.

Process renaissance script relies on the way duetbench stores results
i.e. single directory with multiple properly named files, this should be
extensible for other types of benchmarks as well since the naming
scheme will be kept even for multiple files specified in config and the
processing script has access to this config.

Add CI
* github actions python-app workflow
* formatting with black and checks with flake8
* add python duet package
* add pre-commit with formatting hooks
* add tests

Update README.md with guide on how to run Asynchronous-Duet and fix code
so that it runs that way.

Add example result files.

Update package so that it installs process scripts with pip install .[all]